pkg/issuer/dns: use %s and %q for errs and quoting
The golang spec defines %s to call '.Error()' for things fulfilling the
error interface (and then '.String()' if it's a stringer, with less
priority). Because of that, calling `.Error` is redundant.

The '%q' format string also does a slightly better job of quoting than
doing it by hand with %s.